packages/pytess/pyproject.toml
[project]
name = "pytess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "pillow>=11.2.1",
  "pytesseract>=0.3.13",
]

[project.scripts]
pytess = "pytess:main"
validate-setup = "pytess.validate_setup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
